Move iam schema into own file
